Ring of Power
The Ring of Power, based off of the One Ring from Tolkein Lore, is a new
artifact added to the game. The Ring of Power has many special effects
in line with Tolkein Lore, inlcuding appearing on hobbits, viewing the
ring text when in fire (but only readable when you are a wizard),
messages when throwing into lava, and even Nazgul to follow you around
and ruin your day.