Use monster naming functions when drunken boxing.
Instead of directly displaying the monster names when drunken boxing,
instead use the grammatically correct monster naming functions. That
way, you don't give away too much information to the player if they are
e.x. blinded.

For example, `You poke the Wizard of Yendor in the eye for good
measure.` becomes `You poke it in the eye for good measure.`

Fixes #13.