Make the output files of makedefs depend on makedefs.
Previously, there was an implicit dependency on the _presence_ of
makedefs, but not a hard dependency on it. Since game information which
is written out by makedefs (e.x. artifacts) is compiled directly into
the binary, a weak dependency simply on the presence of makedefs is
insufficient. A strong dependency is needed to make the files generated
by makedefs be rebuilt when e.x. artifacts are added to the game.